kufhjfhgfhgf

+++ src/main/java/com/cootramixtol/sipd/repositories/ConductorRepository.java
@@ -12,6 +12,6 @@ public interface ConductorRepository extends JpaRepository<Conductor, Long>{
     
     Conductor findById(long id);
     List<Conductor> findByNombresContains(String nombres);
-    Object findByIdentificacion(int identificacion);
+    Object findByIdentificacion(Long identificacion);
 
 }
